other changes

+++ sys/drv/vga/include/video.h
@@ -10,7 +10,7 @@ s32 unused =			0;
 s8 *vga_vidmem =		(s8 *)0xb8000;
 s8 vga_unused =			0;
 
-void vga_write(s8 *, s32, s8);
+void vga_write(s8 *, s16, s8);
 void vga_clearscreen(void);
 
 #endif
